refactor: change to v1 for eventual v2 api
