+++ app/Http/Requests/v1/StoreLoadOrderRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\v1;
 
 use App\Rules\ValidFilename;
 use App\Rules\ValidMimeType;
